cleanup storage code
