+++ openpathsampling/netcdfplus/__init__.py
@@ -1,10 +1,14 @@
-from netcdfplus import NetCDFPlus
 from base import StorableNamedObject, StorableObject, create_to_dict
-from proxy import DelayedLoader, lazy_loading_attributes, LoaderProxy
 from cache import WeakKeyCache, WeakLRUCache, WeakValueCache, MaxCache, \
     NoCache, Cache, LRUCache, LRUChunkLoadingCache
 from dictify import ObjectJSON, StorableObjectJSON, UUIDObjectJSON
-from objects import ObjectStore, VariableStore, DictStore, NamedObjectStore, \
-    UniqueNamedObjectStore, ImmutableDictStore
+from netcdfplus import NetCDFPlus
 
+from stores import ObjectStore
+from stores import IndexedObjectStore
+from stores import VariableStore
+from stores import DictStore, ImmutableDictStore
+from stores import NamedObjectStore, UniqueNamedObjectStore
+
+from proxy import DelayedLoader, lazy_loading_attributes, LoaderProxy
 from util import with_timing_logging
