+++ openpathsampling/netcdfplus/stores/object.py
@@ -1,11 +1,10 @@
 import logging
 from uuid import UUID
-
-from cache import MaxCache, Cache, NoCache, WeakLRUCache
-from proxy import LoaderProxy
-from base import StorableNamedObject, StorableObject
-
 from weakref import WeakValueDictionary
+
+from openpathsampling.netcdfplus.base import StorableNamedObject, StorableObject
+from openpathsampling.netcdfplus.cache import MaxCache, Cache, NoCache, WeakLRUCache
+from openpathsampling.netcdfplus.proxy import LoaderProxy
 
 logger = logging.getLogger(__name__)
 init_log = logging.getLogger('openpathsampling.initialization')
@@ -918,837 +917,3 @@ class ObjectStore(StorableNamedObject):
 
     def _get_id(self, idx, obj):
         obj.__uuid__ = self.vars['uuid'][idx]
-
-
-class NamedObjectStore(ObjectStore):
-    def __init__(self, content_class, json=True, nestable=False):
-        super(NamedObjectStore, self).__init__(
-            content_class=content_class,
-            json=json,
-            nestable=nestable
-        )
-
-        self._names_loaded = False
-        self._name_idx = dict()
-
-        if self.content_class is not None \
-                and not issubclass(self.content_class, StorableNamedObject):
-            raise ValueError((
-                'Content class "%s" must be subclassed from '
-                'StorableNamedObject.') %
-                self.content_class.__name__
-            )
-
-    def initialize(self):
-        """
-        Initialize the associated storage to allow for object storage. Mainly
-        creates an index dimension with the name of the object.
-        """
-        super(NamedObjectStore, self).initialize()
-
-        self.create_variable(
-            "name", 'str',
-            description='The name of the object',
-            chunksizes=tuple([65536])
-        )
-
-    def add_single_to_cache(self, idx, json):
-        """
-        Add a single object to cache by json
-
-        Parameters
-        ----------
-        idx : int
-            the index where the object was stored
-        json : str
-            json string the represents a serialized version of the stored object
-        """
-
-        if idx not in self.cache:
-            obj = super(NamedObjectStore, self).add_single_to_cache(idx, json)
-
-            name = self.storage.variables[self.prefix + '_name'][idx]
-            setattr(obj, '_name', name)
-            if name != '':
-                self._update_name_in_cache(obj._name, idx)
-
-    @property
-    def name_idx(self):
-        """
-        Returns a dictionary of all names pointing to stored indices
-
-        Returns
-        -------
-        dict of str : set
-            A dictionary that has all stored names as keys and the values are a
-            set of indices where an object with this name is found.
-
-        """
-
-        # if not done already cache names once
-        if not self._names_loaded:
-            self.update_name_cache()
-
-        return self._name_idx
-
-    def update_name_cache(self):
-        """
-        Update the internal name cache with all stored names in the store.
-
-        This allows to load by name for named objects
-        """
-        if not self._names_loaded:
-            for idx, name in enumerate(
-                    self.storage.variables[self.prefix + "_name"][:]):
-                self._update_name_in_cache(name, idx)
-
-            self._names_loaded = True
-
-    def _update_name_in_cache(self, name, idx):
-        # make sure to cast unicode to str
-        name = str(name)
-        if name != '':
-            if name not in self._name_idx:
-                self._name_idx[name] = {idx}
-            else:
-                if idx not in self._name_idx[name]:
-                    self._name_idx[name].add(idx)
-
-    def find(self, name):
-        """
-        Return last object with a given name
-
-        Parameters
-        ----------
-        name : str
-            the name to be searched for
-
-        Returns
-        -------
-        :py:class:`openpathsampling.netcdfplus.base.StorableObject`
-            the last object with a given name. This is to mimic immutable
-            objects. Once you (re-)save with the same name you replace the
-            old one and hence you leed to load the last stored one.
-
-        """
-        return self.load(name)
-
-    def find_indices(self, name):
-        """
-        Return indices for all objects with a given name
-
-        Parameters
-        ----------
-        name : str
-            the name to be searched for
-
-        Returns
-        -------
-        list of int
-            a list of indices in the storage for all found objects,
-            can be empty [] if no objects with that name exist
-
-        """
-        return sorted(list(self.name_idx[name]))
-
-    def find_all(self, name):
-        if len(self.name_idx[name]) > 0:
-            return self[sorted(list(self.name_idx[name]))]
-
-    # ==========================================================================
-    # LOAD/SAVE DECORATORS FOR CACHE HANDLING
-    # ==========================================================================
-
-    def load(self, idx):
-        """
-        Returns an object from the storage.
-
-        Parameters
-        ----------
-        idx : int or str
-            either the integer index of the object to be loaded or a string
-            (name) for named objects. This will always return the last object
-            found with the specified name. This allows to effectively change
-            existing objects.
-
-        Returns
-        -------
-        :py:class:`openpathsampling.netcdfplus.base.StorableNamedObject`
-            the loaded object
-        """
-
-        if type(idx) is int and idx < 0:
-            return None
-
-        n_idx = idx
-
-        if type(idx) is str:
-            # we want to load by name and it was not in cache.
-            if idx in self.name_idx:
-                if len(self.name_idx[idx]) > 1:
-                    logger.debug((
-                        'Found name "%s" multiple (%d) times in storage! '
-                        'Loading last!') % (
-                        idx, len(self.cache[idx])))
-
-                n_idx = sorted(list(self.name_idx[idx]))[-1]
-            else:
-                raise ValueError('str "' + idx + '" not found in storage')
-
-        elif type(idx) is long:
-            pass
-
-        elif type(idx) is not int:
-            raise ValueError((
-                'indices of type "%s" are not allowed in named '
-                'storage (only str and int)') %
-                type(idx).__name__
-            )
-
-        obj = super(NamedObjectStore, self).load(n_idx)
-
-        if obj is not None:
-            n_idx = self.index[obj.__uuid__]
-            setattr(obj, '_name',
-                    self.storage.variables[self.prefix + '_name'][n_idx])
-            # make sure that you cannot change the name of loaded objects
-            obj.fix_name()
-
-            # finally store the name of a named object in cache
-            self._update_name_in_cache(obj._name, n_idx)
-
-        return obj
-
-    def save(self, obj, idx=None):
-        """
-        Saves an object to the storage.
-
-        Parameters
-        ----------
-        obj : :py:class:`openpathsampling.netcdfplus.base.StorableNamedObject`
-            the object to be stored
-        idx : int or string or `None`
-            the index to be used for storing. This is highly discouraged since
-            it changes an immutable object (at least in the storage). It is
-            better to store also the new object and just ignore the
-            previously stored one.
-
-        """
-
-        is_str = type(idx) is str
-
-        if not is_str and idx is not None:
-            raise ValueError(
-                'Unsupported index type (only str or None allowed).')
-
-        name = obj._name
-
-        if is_str:
-            obj.name = idx
-            name = obj._name
-
-        if name is None:
-            # this should not happen!
-            logger.debug(
-                "Nameable object has not been initialized correctly. "
-                "Has None in _name")
-            raise AttributeError(
-                '_name needs to be a string for nameable objects.')
-
-        obj_fixed = obj._name_fixed
-        obj_name = obj._name
-        # we fix the name just in case we try in recursive saving to store
-        # one object twice with different names. Storing with the same name
-        # is fine!
-        obj.fix_name()
-
-        try:
-            reference = super(NamedObjectStore, self).save(obj)
-        except:
-            # if saving did not work unlock the name if is was un-fixed before
-            obj._name_fixed = obj_fixed
-            obj._name = obj_name
-            raise
-
-        n_idx = self.index[obj.__uuid__]
-        self.storage.variables[self.prefix + '_name'][n_idx] = name
-        self._update_name_in_cache(name, n_idx)
-
-        return reference
-
-
-class UniqueNamedObjectStore(NamedObjectStore):
-
-    # ==========================================================================
-    # LOAD/SAVE DECORATORS FOR CACHE HANDLING
-    # ==========================================================================
-
-    def __init__(self, content_class, json=True, nestable=False):
-        super(UniqueNamedObjectStore, self).__init__(
-            content_class=content_class,
-            json=json,
-            nestable=nestable)
-
-        self._free_name = set()
-
-    def reserve_name(self, name):
-        """
-        Locks a name as used
-
-        Parameters
-        ----------
-        name : str
-            the name to be locked for storage
-        """
-        if name != "":
-            self._free_name.add(name)
-
-    def release_name(self, name):
-        """
-        Releases a locked name
-
-        Parameters
-        ----------
-        name : str
-            the name to be released for being used as a name
-        """
-        self._free_name.discard(name)
-
-    def is_name_locked(self, name):
-        """
-        Test whether in a unique name store a name is already taken
-
-        Parameters
-        ----------
-        name : str or `None`
-            the name to be tested.
-
-        Returns
-        -------
-        bool
-            the result of the test. If the name exists or is reserved during
-            a saving event this will return `True` and return `False` if the
-            name is free.
-
-        """
-        if name is None:
-            return False
-
-        return name in self.name_idx or name in self._free_name
-
-    def save(self, obj, idx=None):
-        """
-        Saves an object to the storage.
-
-        Parameters
-        ----------
-        obj : :py:class:`openpathsampling.netcdfplus.base.StorableNamedObject`
-            the object to be stored
-        idx : string or `None`
-            the index to be used for storing. This is highly discouraged since
-            it changes an immutable object (at least in the storage). It is
-            better to store also the new object and just ignore the
-            previously stored one.
-
-        """
-
-        is_str = type(idx) is str
-
-        if not is_str and idx is not None:
-            raise ValueError(
-                'Unsupported index type (only str or None allowed).')
-
-        name = obj._name
-        fixed = obj._name_fixed
-        err = list()
-
-        if is_str:
-            if fixed:
-                if name != idx:
-                    # saving fixed under different name is not possible.
-                    if obj in self:
-                        err.append(
-                            ('Cannot rename object to "%s". '
-                             'Already saved with name "%s" !') % (idx, name)
-                        )
-                    else:
-                        err.append(
-                            ('Cannot rename object to "%s". '
-                             'Already fixed name "%s" !') % (idx, name)
-                        )
-
-                        if self.is_name_locked(name):
-                            err.append(
-                                ('Current name "%s" is also already taken in '
-                                 'unique name store. This means you cannot '
-                                 'save object "%s" at all. '
-                                 'In general this should not happen to unsaved '
-                                 'objects unless you fixed the name of the '
-                                 'object yourself. Check your code '
-                                 'for the generation of objects of the same '
-                                 'name.') %
-                                (name, obj)
-                            )
-                        else:
-                            err.append(
-                                ('Current name "%s" is still free. Saving '
-                                 'without giving a specific name '
-                                 'should work. If that is what you want '
-                                 'to do.') % name
-                            )
-                else:
-                    # already fixed, but with same name. Okay. Check if stored
-                    if obj in self.index:
-                        return self.reference(obj)
-            else:
-                # name is not fixed yet. Check, if we can save or name is taken
-                if self.is_name_locked(idx):
-                    err.append(
-                        ('New name "%s" already taken in unique name store. ' +
-                         'Try different name instead.') % idx
-                    )
-
-                    if self.is_name_locked(name):
-                        err.append((
-                            'Current name "%s" already taken in unique '
-                            'name store. ') % name
-                        )
-                    else:
-                        err.append(
-                            ('Current name "%s" is still free. Saving without '
-                             'giving a specific name should work') % name
-                        )
-        else:
-            if fixed:
-                # no new name, but fixed. Check if already stored.
-                if obj.__uuid__ in self.index:
-                    return self.reference(obj)
-
-                # if not stored yet check if we could
-                if self.is_name_locked(name):
-                    err.append(
-                        ('Current name "%s" is already taken in unique name '
-                         'store. This means you cannot save object "%s" at '
-                         'all. In general this should not happen to unsaved '
-                         'objects unless you fixed the name of the object '
-                         'yourself. Check your code for the generation of '
-                         'objects of the same name.') %
-                        (name, obj)
-                    )
-            else:
-                # no new name and not fixed. Just check if current name is taken
-                if self.is_name_locked(name):
-                    err.append(
-                        ('Current name "%s" is already taken in unique name '
-                         'store %s. Try renaming object or saving using other '
-                         'name.') % (name, self.name)
-                    )
-
-        if len(err) > 0:
-            raise RuntimeWarning('/n'.join(err))
-
-        # no errors, reserve the name for nested saving and actually call save
-        self.reserve_name(name)
-
-        try:
-            reference = super(UniqueNamedObjectStore, self).save(obj, idx)
-        finally:
-            self.release_name(name)
-
-        return reference
-
-
-class VariableStore(ObjectStore):
-    def __init__(self, content_class, var_names):
-        super(VariableStore, self).__init__(
-            content_class,
-            json=False
-        )
-
-        # TODO: determine var_names automatically from content_class
-        # problem is that some decorators, e.g. using delayed loader
-        # hide the actual __init__ signature and so we cannot determine
-        # what variables to store. Could be 2.0
-
-        if not issubclass(content_class, StorableObject):
-            raise ValueError(('Content_class %s must be subclassed from '
-                             'StorableObject') % content_class.__name__)
-
-        self.var_names = var_names
-        self._cached_all = False
-
-    def to_dict(self):
-        return {
-            'content_class': self.content_class,
-            'var_names': self.var_names
-        }
-
-    def _save(self, obj, idx):
-        for var in self.var_names:
-            self.write(var, idx, obj)
-
-    def _load(self, idx):
-        # attr = {var: self.vars[var][idx] for var in self.var_names}
-        args = [ self.vars[var][idx] for var in self.var_names]
-        return self.content_class(*args)
-
-    def initialize(self):
-        super(VariableStore, self).initialize()
-
-        # Add here the stores to be imported
-        # self.create_variable('name', 'var_type')
-
-    def all(self):
-        self.cache_all()
-        return self
-
-    def cache_all(self, part=None):
-        """Load all samples as fast as possible into the cache
-
-        Parameters
-        ----------
-        part : list of int or `None`
-            If `None` (default) all samples will be loaded. Otherwise the
-            list of indices in `part` will be loaded into the cache
-
-        """
-        max_length = self.cache.size[0]
-        max_length = len(self) if max_length < 0 else max_length
-
-        if part is None:
-            length = min(len(self), max_length)
-            part = range(length)
-        else:
-            part = sorted(list(set(part())))
-            length = min(len(part), max_length)
-            part = part[:length]
-
-        if not part:
-            return
-
-        # just in case we saved the var_names in another order and so we are
-        # backwards compatible
-        var_names = self.content_class.args()[1:]
-
-        if not self._cached_all:
-            data = zip(*[
-                self.vars[var][part]
-                for var in var_names
-            ])
-
-            [self.add_to_cache(idx, v) for idx, v in zip(part, data)]
-
-            self._cached_all = True
-
-    def add_to_cache(self, idx, data):
-        if idx not in self.cache:
-            # attr = {var: self.vars[var].getter(data[nn])
-            #         for nn, var in enumerate(self.var_names)}
-            obj = self.content_class(*data)
-            self._get_id(idx, obj)
-
-            self.index[obj.__uuid__] = idx
-            self.cache[idx] = obj
-
-
-class DictStore(NamedObjectStore):
-    def __init__(self):
-        super(DictStore, self).__init__(
-            None,
-            json='json'
-        )
-
-    def to_dict(self):
-        return {}
-
-    def load(self, idx):
-        """
-        Returns an object from the storage.
-
-        Parameters
-        ----------
-        idx : str
-            a string (name) of the objects. This will always return the
-            last object found with the specified name. If immutable is true
-            for the store it assures that there is only a single object per name
-
-        Returns
-        -------
-        :class:`openpathsampling.netcdfplus.base.StorableObject`
-            the loaded object
-        """
-
-        if type(idx) is str:
-            n_idx = -1
-
-            # we want to load by name and it was not in cache.
-            if idx not in self.name_idx:
-                logger.debug('Name "%s" not found in the storage!' % idx)
-                raise KeyError('str "' + idx + '" not found in storage')
-
-            if idx in self.name_idx:
-                if len(self.name_idx[idx]) > 1:
-                    logger.debug((
-                        'Found name "%s" multiple (%d) times in storage! '
-                        'Loading last!') % (
-                        idx, len(self.name_idx[idx])))
-
-                n_idx = sorted(list(self.name_idx[idx]))[-1]
-
-                logger.debug(
-                    'Found name "%s" in storage! Loading IDX %d!' %
-                    (idx, n_idx))
-
-        elif type(idx) is int:
-            n_idx = idx
-        else:
-            raise ValueError(
-                'Unsupported index type (only str and int allowed).')
-
-        # turn into python int if it was a numpy int (in some rare cases!)
-        n_idx = int(n_idx)
-
-        logger.debug(
-            'Calling load object of type `%s` and @ IDX #%d' %
-            (str(self.content_class), n_idx))
-
-        if n_idx >= len(self):
-            logger.warning(
-                'Trying to load from IDX #' + str(n_idx) +
-                ' > number of objects ' + str(len(self))
-            )
-            raise RuntimeError(
-                'Loading of too large int should be attempted. '
-                'Problem in name cache. This should never happen!')
-        elif n_idx < 0:
-            logger.warning(
-                'Trying to load negative IDX #' + str(n_idx) + ' < 0. '
-                'This should never happen!!!'
-            )
-            raise RuntimeError(
-                'Loading of negative int should result in no object. '
-                'This should never happen!'
-            )
-        else:
-
-            logger.debug('Loading named object from index IDX # %d' % n_idx)
-            obj = self._load(n_idx)
-
-            logger.debug(
-                'Loading named object from index IDX # %d.. DONE' % n_idx)
-
-        return obj
-
-    def restore(self):
-        pass
-
-    def save(self, obj, idx=None):
-        """
-        Saves an object to the storage.
-
-        Parameters
-        ----------
-        obj : :py:class:`openpathsampling.netcdfplus.base.StorableObject`
-            the object to be stored
-        idx : string or `None`
-            the string index to be used for storing. Objects will not be
-            replaced but stored again with the same name. When loading the
-            last stored object under the idx is retrieved. Effectively mimicking
-            a mutual dict with versioning. We usually encourage for most cases
-            to use the immutual dict class
-            :class:`openpathsampling.netcdfplus.ImmutableDictStore` instead
-            to avoid ambiguity in stored objects.
-
-        See Also
-        --------
-        :class:`openpathsampling.netcdfplus.ImmutableDictStore`
-
-        """
-
-        if idx is None:
-            # a DictStore needs a specific name
-            raise ValueError(
-                'Saving in a DictStore without specifying a '
-                'string key is not allowed. ')
-
-        if type(idx) is not str:
-            # key needs to be a string
-            raise ValueError(
-                'Index `%s` for DictStore needs to be a string! ' % idx)
-
-        n_idx = int(self.free())
-        # make sure in nested saving that an IDX is not used twice!
-        self.reserve_idx(n_idx)
-
-        logger.debug(
-            'Saving `%s` with name `%s` @ IDX #%d' %
-            (str(obj.__class__), idx, n_idx))
-        self._save(obj, n_idx)
-
-        self.storage.variables[self.prefix + '_name'][n_idx] = idx
-        self._update_name_in_cache(idx, n_idx)
-
-        return n_idx
-
-    def keys(self):
-        return self.name_idx.keys()
-
-    def iterkeys(self):
-        return self.name_idx.iterkeys()
-
-    def __iter__(self):
-        return self.iterkeys()
-
-    def iteritems(self):
-        for name in self:
-            yield name, self[name]
-
-    def get(self, idx, default=None):
-        try:
-            return self.load(idx)
-        except KeyError:
-            return default
-
-
-class ImmutableDictStore(DictStore):
-
-    def save(self, obj, idx=None):
-        """
-        Saves an object to the storage.
-
-        Parameters
-        ----------
-        obj : :py:class:`openpathsampling.netcdfplus.base.StorableObject`
-            the object to be stored
-        idx : int or string or `None`
-            the index to be used for storing. This is highly discouraged since
-            it changes an immutable object (at least in the storage). It is
-            better to store also the new object and just ignore the
-            previously stored one.
-
-        """
-
-        if idx in self.name_idx:
-            # immutable means no duplicates, so quit
-            raise RuntimeWarning(
-                'Cannot re-save existing key "%s" in '
-                'immutable dict store.' % idx
-            )
-
-        return super(ImmutableDictStore, self).save(obj, idx)
-
-
-class IndexedObjectStore(ObjectStore):
-    """
-    ObjectStore storing objects in arbitrary order
-
-    This has a prefilled .index which knows at which position a certain
-    index is stored. This way you can circumvent holes and keep the file smaller
-    """
-
-    # ==========================================================================
-    # LOAD/SAVE DECORATORS FOR CACHE HANDLING
-    # ==========================================================================
-
-    def load(self, idx):
-        """
-        Returns an object from the storage.
-
-        Parameters
-        ----------
-        idx : int
-            the integer index of the object to be loaded
-
-        Returns
-        -------
-        :py:class:`openpathsampling.netcdfplus.base.StorableObject`
-            the loaded object
-        """
-
-        # we want to load by uuid and it was not in cache.
-        if idx in self.index:
-            n_idx = self.index[idx]
-        else:
-            raise KeyError(idx)
-
-        if n_idx < 0:
-            return None
-
-        # if it is in the cache, return it
-        try:
-            obj = self.cache[n_idx]
-            return obj
-
-        except KeyError:
-            pass
-
-        obj = self._load(n_idx)
-        self.cache[n_idx] = obj
-
-        return obj
-
-    # def create_uuid_index(self):
-    #     return dict()
-
-    def save(self, obj, idx=None):
-        """
-        Saves an object to the storage.
-
-        Parameters
-        ----------
-        obj : :py:class:`openpathsampling.netcdfplus.base.StorableObject`
-            the object to be stored
-        idx : int or string or `None`
-            the index to be used for storing. This is highly discouraged since
-            it changes an immutable object (at least in the storage). It is
-            better to store also the new object and just ignore the
-            previously stored one.
-
-        """
-
-        if idx in self.index:
-            # has been saved so quit and do nothing
-            return idx
-
-        # n_idx = self.free()
-        n_idx = len(self.index)
-
-        # mark as saved so circular dependencies will not cause infinite loops
-        self.index.append(idx)
-
-        # make sure in nested saving that an IDX is not used twice!
-        # self.reserve_idx(n_idx)
-
-        logger.debug('Saving ' + str(type(obj)) + ' using IDX #' + str(n_idx))
-
-        try:
-            self._save(obj, n_idx)
-            self.vars['index'][n_idx] = idx
-
-            # store the name in the cache
-            # if hasattr(self, 'cache'):
-            self.cache[n_idx] = obj
-
-        except:
-            logger.debug('Problem saving %d !' % n_idx)
-            # in case we did not succeed remove the mark as being saved
-            del self.index[idx]
-            # self.release_idx(n_idx)
-            raise
-
-        # self.release_idx(n_idx)
-        self._set_id(n_idx, obj)
-
-        return idx
-
-    def restore(self):
-        self.index.clear()
-        self.index.extend(self.vars['index'][:])
-
-    def initialize(self):
-        super(IndexedObjectStore, self).initialize()
-
-        self.create_variable(
-            'index',
-            'index'
-        )
